// apps/admin-frontend/services/interfaces/createMedicalRecordRequestDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * TrustMed Core API
 * API documentation for the TrustMed Core backend
 * OpenAPI spec version: 1.0.0
 */

export interface CreateMedicalRecordRequestDto {
  file: Blob;
  personId: string;
  category: string;
  notes?: string;
  doctorName?: string;
  hospitalName?: string;
  recordDate?: string;
}