// apps/admin-frontend/services/interfaces/updateMedicalRecordRequestDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * TrustMed Core API
 * API documentation for the TrustMed Core backend
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateMedicalRecordRequestDtoCategory } from './updateMedicalRecordRequestDtoCategory';

export interface UpdateMedicalRecordRequestDto {
  category?: UpdateMedicalRecordRequestDtoCategory;
  notes?: string;
  doctorName?: string;
  hospitalName?: string;
  recordDate?: string;
}